Fix imports, use uv instead of poetry, add docker publish in CI
