+++ .gitignore
@@ -169,3 +169,5 @@ cython_debug/
 #  option (not recommended) you can uncomment the following to ignore the entire idea folder.
 #.idea/
 .history
+.vscode
+gpmfstream/
