+++ Dockerfile
@@ -1,30 +1,21 @@
-FROM python:3.10-slim
+FROM nvidia/cuda:11.8.0-base-ubuntu22.04
 
 # Set commit hash for gpmfstream repo as there are no versions
 ARG GPMFSTREAM_GIT_HASH=f1a9742
-
-RUN pip install poetry pybind11
-
-WORKDIR /app
+ARG DEBIAN_FRONTEND=noninteractive
 
 # # Get the gpmfstream library
 RUN apt-get update && \
-    apt-get install -y git wget unzip build-essential \
-    curl ffmpeg libsm6 libxext6
+    apt-get install -y git ffmpeg pipx python3-pip \
+    && pip3 install uv
 
-RUN git clone https://github.com/hovren/gpmfstream.git
-WORKDIR /app/gpmfstream
-RUN git checkout ${GPMFSTREAM_GIT_HASH} && git submodule update --init && python setup.py install
-
+# RUN pip3 install poetry
 WORKDIR /app
 
-# Copy the model checkpoints
+# Copy the model checkpoints and environment
 COPY segmentation_net.pth /app/
 COPY sfm_net.pth /app/
-
-# Copy the pyproject.toml and poetry.lock files to the container
-COPY poetry.lock pyproject.toml /app/
-
+COPY uv.lock pyproject.toml /app/
 
 # Change to non-root user
 RUN groupadd mygroup --gid 1000 && \
@@ -37,12 +28,22 @@ RUN groupadd mygroup --gid 1000 && \
 
 USER 1000
 
-# Install the dependencies
-RUN poetry install --no-root --no-dev
+# Build gpmfstream
+RUN git clone https://github.com/hovren/gpmfstream.git
+WORKDIR /app/gpmfstream
+RUN pip3 install pybind11 setuptools
+RUN git checkout ${GPMFSTREAM_GIT_HASH} \
+    && git submodule update --init
+RUN python3 setup.py bdist_wheel
+
+# Install dependencies
+WORKDIR /app
+RUN uv sync --no-dev --no-cache --extra build --extra compile
+RUN uv pip install "./gpmfstream/dist/gpmfstream-0.5-cp310-cp310-linux_x86_64.whl"
 
 COPY src /app/src
 COPY example_inputs /app/example_inputs
 
 WORKDIR /app/src
 
-ENTRYPOINT ["poetry", "run", "python3", "reconstruct.py"]
+ENTRYPOINT ["uv", "run", "python3", "reconstruct.py"]
